Merging simulation to main

+++ Core/Inc/lookuptable.h
@@ -1,11 +1,11 @@
 #define LENGTH_LUT 100
 
-extern int currentVectorX[];
-extern int currentVectorY[];
-extern int currentVectorZ[];
-extern int angularVelVectorX[];
-extern int angularVelVectorY[];
-extern int angularVelVectorZ[];
-extern int positionVectorX[];
-extern int positionVectorY[];
-extern int positionVectorZ[];
+extern float currentVectorX[];
+extern float currentVectorY[];
+extern float currentVectorZ[];
+extern float angularVelVectorX[];
+extern float angularVelVectorY[];
+extern float angularVelVectorZ[];
+extern float positionVectorX[];
+extern float positionVectorY[];
+extern float positionVectorZ[];
